requirements.txt
flask
requests
gunicorn
pyyaml

